move into folder and add expansion span
