+++ src/sass/_typography.scss
@@ -14,3 +14,16 @@
   font-weight: 600;
   line-height: 24px;
 }
+
+@mixin summary-count-text() {
+  color: settings.$color-primary;
+  font-family: settings.$font-family-primary;
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 24px;
+
+  &:before {
+    color: settings.$color-text;
+    content: ', ';
+  }
+}
